Revert "Merge branch 'main' into coming-soon-pages"
This reverts commit c64ca3f6cc2ed0fc39d3327c38bd0616f5f1ce46, reversing
changes made to 5e57f282b43a6280a4e6be8fdebbd05466deb14d.

+++ packages/kbn-esql-ast/src/visitor/contexts.ts
@@ -337,7 +337,7 @@ export class LimitCommandVisitorContext<
     if (
       arg &&
       arg.type === 'literal' &&
-      (arg.literalType === 'integer' || arg.literalType === 'double')
+      (arg.literalType === 'integer' || arg.literalType === 'decimal')
     ) {
       return arg;
     }
